# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Task Manager API",
    "version": "1.0.0",
    "description": "REST API for the AI task manager — used by the web app, by agents through the MCP server, and by scripts.\n\n**Identity is attribution, not authentication.** Every request may send an `X-Actor` header (`agent:<name>` or `human:<name>`); it is recorded on every write and defaults to `human:anonymous`. Nothing verifies it — it exists so the timeline can tell an agent's change from a human's, and so agents can be kept from overwriting each other's claimed work.\n\n**Optional access gate.** A self-hosted server may set `API_TOKEN`; every `/api/v1` request must then carry `Authorization: Bearer <token>` (the `bearerAuth` scheme). Without it the API is open — right for localhost, wrong for anything reachable from a network you do not control.\n\n- Status changes go through `POST /api/v1/tasks/{taskId}/transition`, never PATCH. What a status requires is the shape of its payload.\n- Writes that accept `expectedVersion` fail with `VERSION_CONFLICT` when the task changed since it was read.\n- List responses are enveloped as `{ data, meta }`; single resources are returned directly. The events feed is cursor-paged instead.\n- Failures always use the `ErrorResponse` envelope with a `SCREAMING_SNAKE` `code`. Branch on the code, not on the status alone.\n- Dates are ISO 8601 UTC strings.\n- Task ids are also task numbers: task `42` renders as `TASK-000042`."
  },
  "servers": [
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Only when the server sets API_TOKEN: every /api/v1 request must then send `Authorization: Bearer <API_TOKEN>`, or it gets 401 UNAUTHORIZED. /health and /docs stay open. It is one shared secret for every caller — who did what is still the X-Actor header."
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "VALIDATION_ERROR",
                  "AT_LEAST_ONE_FIELD",
                  "UNAUTHORIZED",
                  "ACTOR_NOT_PERMITTED",
                  "TASK_NOT_FOUND",
                  "COMMENT_NOT_FOUND",
                  "VERSION_CONFLICT",
                  "TASK_ALREADY_CLAIMED",
                  "NOT_CLAIM_HOLDER",
                  "DEPENDENCY_CYCLE",
                  "NO_OPEN_DECISION",
                  "MALFORMED_JSON",
                  "PAYLOAD_TOO_LARGE",
                  "NOT_FOUND",
                  "INTERNAL_ERROR"
                ]
              },
              "message": {
                "type": "string",
                "minLength": 1
              },
              "details": {},
              "requestId": {
                "type": "string",
                "minLength": 1
              }
            },
            "required": [
              "code",
              "message",
              "requestId"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false,
        "description": "The error envelope. Clients branch on error.code, never on the HTTP status alone."
      },
...
      "Comment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "exclusiveMinimum": 0
          },
          "taskId": {
            "type": "integer",
            "exclusiveMinimum": 0
          },
          "author": {
            "type": "string",
            "minLength": 1
          },
          "kind": {
            "type": "string",
            "enum": [
              "note",
              "progress",
              "qa_feedback"
            ]
          },
          "body": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "taskId",
          "author",
          "kind",
          "body",
          "createdAt"
        ],
        "additionalProperties": false,
        "description": "A comment on a task. Comments are append-only, so there is no updatedAt. `author` is the X-Actor that posted it."
      },
...
    },
...
  },
  "paths": {
...
    "/api/v1/tasks/{taskId}/comments": {
      "post": {
        "tags": [
          "Comments"
        ],
        "summary": "Add a comment",
        "description": "`author` is the X-Actor header. Comments are open to every actor, even on a task another agent has claimed — and a comment from the claim holder renews its lease. Otherwise adding a comment touches neither the task's version nor its updatedAt. The body is validated **before** the parent task is looked up, so a bad payload against a missing task is a 422 rather than a 404.",
        "security": [
          {
            "bearerAuth": []
          },
          {}
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d{1,15}$",
              "description": "Id of the parent task.",
              "example": 42
            },
            "required": true,
            "description": "Id of the parent task.",
            "name": "taskId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "description": "Who is acting: `agent:<name>` or `human:<name>` (case-insensitive; stored lowercase). Recorded on every write — createdBy, comment author, event actor, claim holder. Absent means `human:anonymous`. Malformed is a 422 VALIDATION_ERROR with details[\"X-Actor\"]. `system:` is reserved for the server.",
              "example": "agent:claude-code"
            },
            "required": false,
            "description": "Who is acting: `agent:<name>` or `human:<name>` (case-insensitive; stored lowercase). Recorded on every write — createdBy, comment author, event actor, claim holder. Absent means `human:anonymous`. Malformed is a 422 VALIDATION_ERROR with details[\"X-Actor\"]. `system:` is reserved for the server.",
            "name": "X-Actor",
            "in": "header"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "body": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 5000
                  },
                  "kind": {
                    "type": "string",
                    "enum": [
                      "note",
                      "progress",
                      "qa_feedback"
                    ],
                    "default": "note"
                  }
                },
                "required": [
                  "body"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created. The Location header points at the new comment.",
            "headers": {
              "Location": {
                "description": "URL of the created comment.",
                "schema": {
                  "type": "string",
                  "example": "/api/v1/tasks/42/comments/191"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comment"
                }
              }
            }
          },
          "400": {
            "description": "The request body is not valid JSON.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "MALFORMED_JSON"
                          ]
                        },
                        "message": {
                          "type": "string",
                          "minLength": 1
                        },
                        "details": {},
                        "requestId": {
                          "type": "string",
                          "minLength": 1
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "The server runs with API_TOKEN set and the request carried no `Authorization: Bearer <token>`, or the wrong one. Never returned by a server without API_TOKEN.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "UNAUTHORIZED"
                          ]
                        },
                        "message": {
                          "type": "string",
                          "minLength": 1
                        },
                        "details": {},
                        "requestId": {
                          "type": "string",
                          "minLength": 1
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "No such parent task. Checked explicitly rather than left to the foreign key, which would surface as a 500.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "TASK_NOT_FOUND"
                          ]
                        },
                        "message": {
                          "type": "string",
                          "minLength": 1
                        },
                        "details": {},
                        "requestId": {
                          "type": "string",
                          "minLength": 1
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "413": {
            "description": "The request body is larger than BODY_LIMIT.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "PAYLOAD_TOO_LARGE"
                          ]
                        },
                        "message": {
                          "type": "string",
                          "minLength": 1
                        },
                        "details": {},
                        "requestId": {
                          "type": "string",
                          "minLength": 1
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "422": {
            "description": "A field failed validation. `details` maps field name to messages. A malformed X-Actor header is also a VALIDATION_ERROR, with details[\"X-Actor\"].",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR"
                          ]
                        },
                        "message": {
                          "type": "string",
                          "minLength": 1
                        },
                        "details": {},
                        "requestId": {
                          "type": "string",
                          "minLength": 1
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error. The body carries a requestId and never a stack trace.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "INTERNAL_ERROR"
                          ]
                        },
                        "message": {
                          "type": "string",
                          "minLength": 1
                        },
                        "details": {},
                        "requestId": {
                          "type": "string",
                          "minLength": 1
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Any other failure. Every error response in this API uses the ErrorResponse envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{taskId}/comments/{commentId}": {
      "delete": {
        "tags": [
          "Comments"
        ],
        "summary": "Delete a comment",
        "description": "Scoped by both ids. A comment belonging to a different task is a 404, identical to one that does not exist — so this path cannot be walked to enumerate another task's comment ids.",
        "security": [
          {
            "bearerAuth": []
          },
          {}
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d{1,15}$",
              "description": "Id of the parent task.",
              "example": 42
            },
            "required": true,
            "description": "Id of the parent task.",
            "name": "taskId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d{1,15}$",
              "description": "Id of the comment, which must belong to this task.",
              "example": 191
            },
            "required": true,
            "description": "Id of the comment, which must belong to this task.",
            "name": "commentId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "description": "Who is acting: `agent:<name>` or `human:<name>` (case-insensitive; stored lowercase). Recorded on every write — createdBy, comment author, event actor, claim holder. Absent means `human:anonymous`. Malformed is a 422 VALIDATION_ERROR with details[\"X-Actor\"]. `system:` is reserved for the server.",
              "example": "agent:claude-code"
            },
            "required": false,
            "description": "Who is acting: `agent:<name>` or `human:<name>` (case-insensitive; stored lowercase). Recorded on every write — createdBy, comment author, event actor, claim holder. Absent means `human:anonymous`. Malformed is a 422 VALIDATION_ERROR with details[\"X-Actor\"]. `system:` is reserved for the server.",
            "name": "X-Actor",
            "in": "header"
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted. No body."
          },
          "401": {
            "description": "The server runs with API_TOKEN set and the request carried no `Authorization: Bearer <token>`, or the wrong one. Never returned by a server without API_TOKEN.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "UNAUTHORIZED"
                          ]
                        },
                        "message": {
                          "type": "string",
                          "minLength": 1
                        },
                        "details": {},
                        "requestId": {
                          "type": "string",
                          "minLength": 1
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "No such comment, the comment belongs to another task, or no such task — all three are the same 404 on purpose.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "COMMENT_NOT_FOUND"
                          ]
                        },
                        "message": {
                          "type": "string",
                          "minLength": 1
                        },
                        "details": {},
                        "requestId": {
                          "type": "string",
                          "minLength": 1
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Only the X-Actor header can fail here. A malformed X-Actor header is also a VALIDATION_ERROR, with details[\"X-Actor\"].",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR"
                          ]
                        },
                        "message": {
                          "type": "string",
                          "minLength": 1
                        },
                        "details": {},
                        "requestId": {
                          "type": "string",
                          "minLength": 1
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error. The body carries a requestId and never a stack trace.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "INTERNAL_ERROR"
                          ]
                        },
                        "message": {
                          "type": "string",
                          "minLength": 1
                        },
                        "details": {},
                        "requestId": {
                          "type": "string",
                          "minLength": 1
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Any other failure. Every error response in this API uses the ErrorResponse envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}